split main() tests into separate class
